requirements.txt
flask
gunicorn
PyPDF2
python-docx
pytesseract
pillow
pandas
openpyxl
requests